refactor: address review comments and split distribution tests
